Create README.md
As a final-year undergraduate student at the University of Colombo
Faculty of Science, I am pursuing a degree in Industrial Statistics and
Mathematical Finance. With a strong passion for data analysis, I am
actively seeking opportunities to apply my skills and knowledge in a
professional setting. My ultimate career goal is to become a Data analyst.
and I am dedicated to continuous learning and growth in this field.